drivers: spi: spi-axi-spi-engine: Increase buffer
The FIFO/Offload memory width is 32-bit.

(cherry picked from commit 4d4e2930068994ab7d351544d0f811f9460d1103)

+++ drivers/spi/spi-axi-spi-engine.c
@@ -337,7 +337,7 @@ int spi_engine_offload_load_msg(struct spi_device *spi,
 	struct spi_transfer *xfer;
 	void __iomem *cmd_addr;
 	void __iomem *sdo_addr;
-	const uint8_t *buf;
+	const uint32_t *buf;
 	unsigned int i, j;
 	size_t size;
